requirements.txt
streamlit
regex
nltk
scikit-learn
pdfplumber
google.generativeai
python-dotenv


